updates URL widget
