+++ arches_component_lab/src/arches_component_lab/widgets/URLWidget/URLWidget.vue
@@ -4,18 +4,18 @@ import URLWidgetViewer from "@/arches_component_lab/widgets/URLWidget/components
 
 import { EDIT, VIEW } from "@/arches_component_lab/widgets/constants.ts";
 
-import type { CardXNodeXWidget } from "@/arches_component_lab/types.ts";
 import type {
-    URLDatatype,
-    WidgetMode,
-} from "@/arches_component_lab/widgets/types.ts";
+    CardXNodeXWidget,
+    NodeData,
+} from "@/arches_component_lab/types.ts";
+import type { WidgetMode } from "@/arches_component_lab/widgets/types.ts";
 
 defineProps<{
     mode: WidgetMode;
     nodeAlias: string;
     graphSlug: string;
     cardXNodeXWidgetData: CardXNodeXWidget;
-    value: URLDatatype | null | undefined;
+    value: NodeData | null | undefined;
 }>();
 
 const emit = defineEmits(["update:isDirty", "update:value"]);
