+++ arches_component_lab/src/arches_component_lab/widgets/URLWidget/components/URLWidgetViewer.vue
@@ -1,20 +1,24 @@
 <script setup lang="ts">
 import { useGettext } from "vue3-gettext";
+import type { NodeData } from "@/arches_component_lab/types.ts";
 import type { URLDatatype } from "@/arches_component_lab/widgets/types.ts";
 
 const { $gettext } = useGettext();
 
-const props = defineProps<{
-    value?: URLDatatype | null | undefined;
+const { value } = defineProps<{
+    value?: NodeData | null | undefined;
 }>();
 </script>
 
 <template>
     <a
-        v-if="props.value?.url"
-        :href="props.value.url"
+        v-if="(value?.interchange_value as URLDatatype)?.url"
+        :href="(value?.interchange_value as URLDatatype)?.url"
     >
-        {{ props.value.url_label || props.value.url }}
+        {{
+            (value?.interchange_value as URLDatatype)?.url_label ||
+            (value?.interchange_value as URLDatatype)?.url
+        }}
     </a>
 
     <span v-else>
