+++ arches_component_lab/src/arches_component_lab/widgets/api.ts
@@ -1,7 +1,5 @@
 import arches from "arches";
 
-import type { ResourceInstanceReference } from "@/arches_component_lab/widgets/types.ts";
-
 // TODO: Remove when full replacement of fetchCardXNodeXWidgetData is implemented
 export const fetchWidgetData = async (graphSlug: string, nodeAlias: string) => {
     const response = await fetch(
@@ -34,51 +32,6 @@ export const fetchNodeData = async (graphSlug: string, nodeAlias: string) => {
     } catch (error) {
         throw new Error((error as Error).message || response.statusText);
     }
-};
-
-export const fetchLists = async (
-    nodeAliases: string[] | undefined = undefined,
-) => {
-    const params = new URLSearchParams();
-    nodeAliases?.forEach((alias) => params.append("node_alias", alias));
-    const response = await fetch(`${arches.urls.controlled_lists}?${params}`);
-    try {
-        const parsed = await response.json();
-        if (response.ok) {
-            return parsed;
-        }
-        throw new Error(parsed.message);
-    } catch (error) {
-        throw new Error((error as Error).message || response.statusText);
-    }
-};
-
-export const fetchRelatableResources = async (
-    graphSlug: string,
-    nodeAlias: string,
-    page: number,
-    filterTerm?: string,
-    initialValues?: ResourceInstanceReference[] | null | undefined,
-) => {
-    const params = new URLSearchParams();
-
-    params.append("page", page.toString());
-    if (filterTerm) {
-        params.append("filter_term", filterTerm);
-    }
-    initialValues?.forEach((initialValue) =>
-        params.append("initialValue", initialValue.resource_id),
-    );
-    const response = await fetch(
-        `${arches.urls.api_relatable_resources(
-            graphSlug,
-            nodeAlias,
-        )}?${params}`,
-    );
-
-    const parsed = await response.json();
-    if (!response.ok) throw new Error(parsed.message || response.statusText);
-    return parsed;
 };
 
 export const fetchCardXNodeXWidgetDataFromNodeGroup = async (
